move UniqueSettings to dtypes.py and fix errors
